Clock ticks since initialization are now printed in kerner logs
Clock ticks since initialization are now printed in kerner logs

Useful for debug.
It's only printed when a log_level is specified. That way it's not printed in early boot console.
Whenever klog_write is called with a specified log_level, sys_gticks() is called and the result is printed