fetching page directory by the current process
The structure of the process has architecture-dependent fields,
such as cr3, fss, pmcs... and since Nanvix will support multiple
architectures, it may not be ideal to mix the fields of distinct
architectures into one structure, I think of 2 solutions:

1) Create #ifdef macros that check architecture and define the
fields according to the architecture.
  or
2) Create a copy of pm.h for each architecture and define the
structure according to each architecture.

Both solutions have their advantages and disadvantages, but there
is still a dependence on these fields with some files and functions
throughout the kernel, such as fork.c, pm.c, sched.c, system calls,
and so on. It may be interesting to address these conflicts before
making any major changes to the process structure in order to reduce
architecture dependency on the kernel.

For now, the use of the CR3 field is sufficient.