fix: buffer_virt/phys overriding page tables
In the x86 implementation, the disk buffers were located
just before the kernel entry point and the address was
hardcoded map in the memory.

It happens that in or1k, the kernel is loaded in the very
beginning of the memory, an that buffer's address were
overriding some entries in the page table, which is not
desirable. To fix this, the buffers are now allocated in
a know place and the address is saved into the appropriate
fields.

Note that this implementation interferes with the x86, due
to the mm.h/.c changes, so, it's also need to properly change
some things (in x86) in order to work as expected.