or1k: using hard-coded cpu clock
Looks like the OpenRISC does not tell us what is the CPU frequency.
Digging around the Linux source, it seems that Linux uses BogoMIPS
and loops_per_jiffy stuff to estimate the clock (maybe I'm wrong).

So, the solution I've found, is to hardcode the clock info into
the nanvix/config.h file. Since we are dealing with architecture
things here, maybe would be good to move the config.h file to
an architecture-dependent folder.