or1k: stdarg for or1k
Since the OpenRISC ABI's is different from x86, the way that variadic
arguments are used is also different.

In order to get around this situation, now the sdarg header is
architecture-dependent and, when in or1k, we use the GCC builtin functions
to accomplish that.

As a side effect, the GCC va_arg builtin, also requires the presence of
the function abort(), which is used in case of fail. So I've made a stub
version of abort() in kerneland that does nothing, just to work as
expected.

Note that this may no to be the best solution, but since kprintf uses
va_args, it's good to have a working version for debug purposes in this
stage of development.