or1k: fix frame pointer, checking present bit when inside xtlb handler
Since the pte entry is created gradually, we need to ensure
that when handling a xTLB miss, the page *really* exists,
even when there are some value in the memory, i.e, we need
to check the present bit for the current pte, to ensure that
the page actually exists. If not, a page vfault is triggered,
which handles for demand fill or demand zero.