or1K: executable pages for everyone
The or1k architecture enables support for executable (or not) pages for
supervisor and user mode separately, which is indeed a very good feature
to have.

Although he this kind of feature (in or1k), the x86 does not make use of
that because, as far as I known, the XD/NX bit is designed for 64-bit
systems only, and so, the Nanvix can only protect pages for read/write
and does not distinguish code pages from data pages.