or1k, bug fix: CoW was not copying the pages
Three issues were responsible to this:
1) Wrong 'p[dt]e_is_write' function: since the function was and-ing the
PPI value to the RW flag, even when we have a read-only page (PPI = 1)
the function returns a non-zero value, since 1 & 3 (RW = 3), is 1. The
fix is quite straightforward: we just have to check if the value is equal
the RW flag.

2) Buggy physcpy function: the physcpy function had two bugs: the function
deactivacted the MMU before to jump to the lower-half kernel (we want the
opposite). Besides that, we were masking only the half-word with the SPR_SR
which could be dangerous, since the higher bits were set to 0.

3) Finally, to work fine, we also need to flush the TLB to reflect the PTE
that has been changed in memory.