temporarily forcing entry point to _start
For unknown reasons, it seems that the linker does not recognize
(or ignores) the ENTRY(_start) line of the lib/linker.ld file.
Due that, the crt0.o file is not included in user-space programs
and the entry-point address is completely wrong.

However, it was noticed that when configuring the entry point
via flags, _start is finally recognized, crt0.o included, and the
entry point address is configured correctly. Therefore, although
redundant, it seems that the '-Wl,-e' flag is our best bet at this
moment.