building ubin, strip/objcpy arch-specific, or1k toolchain..
Here we introduce some things, mainly in relation to the OpenRISC toolchain.
The filesystem structures saved in initrd.img are highly affected by the byte-
ordering of the target architecture. Since the Minix tools (for the host
system) are built on the host itself (x86-64), they end up being little-endian.
OpenRISC, in turn, is big-endian, and therefore, all structures are read 'wrong'.

As a way of getting around, there are two possible solutions:
1) Create a representation of all strucutres used for big/little endian.
2) Switch toolchain: if the initrd.img is also generated using OpenRISC, we have
the guarantee that the generated output is the same as it expects to be read.

This commit introduces the second option. Although the former is indeed the way to
go, it requires that the structures be carefully changed, so the second alternative
is a quick way of having the system working. It should only be noted that this is a
palliative measure and *must* be changed later.