or1k: significant changes to timer interrupt
This commit focuses on two major changes:
- Timer mode is no longer one-shot timer, but auto-restart: we want the
timer to work even when its interruptions is maked, and, to my knowledge,
this is not possible with one-shot mode.

- Most of the issues with the timer were due to poor understanding of the
SPR[TEE] and SPR[IEE] bits. TEE effectively serves to mask timer interrupts
while IEE enables and disables interrupts globally. The crucial point is
that TEE is not directly related to IEE, that is, for a timer exception
to occur, it's only necessary that TEE is active, not both -- as I thought.

So, all the changes made here are to address the points presented above.