or1k: do_hwint() routine
One thing should be noted here: we can't enable interrupts before read the
values available in the serial device, thus, if, and only if, we receive an
IRQ from the serial device, we don't enable interrupts before calling the
handler.

This is not the best approach, indeed, but since we do not have some
mechanism to pass the data into the handler, the quick way is just keep
interrupts disabled.