or1k, bug fix: incoherent values while printing double values
In commit b000f4e a patch was introduced to prevent warnings messages using
the follow idea: whenever va_arg(v,p) is used, type p will be int, so any
type could be passed as an argument as long as the appropriate casting was
performed afterwards.

It turns out that this does not work for all cases, like double for
instance. So when printing double values, a completely wrong value was
displayed, since double in or1k are 8 bytes and int only 4.

The appropriate solution is introduced in this commit: all references to
va_arg() that use non-permitted types (such as char, short, float...) are
then used the equivalent promoted types, such as int and double. The macro
no longer makes castings or anything related, just as should be.