or1k: using shadow registers when available
When an exception occurs, one of the tasks that we have to do is save the
stack. It happens that to accomplish this, we need to save a few registers
beforehand and use them as scratch registers, while we're restoring the
stack and so on...

In the UP implementation, we use the addresses below 0x100 to save registers
and we are safe with that. In the SMP implementation, on the other hand, this
ends up to messing other cores since we are using the same memory region...
The proper solution is the use of shadow registers, although this is an
optional feature, there's no other way to save registers without it, we just
need to ensure that we will have at least one gpr shadow file.

This commit, therefore, adds shadow registers support when the compiler flag
OR1K_HAVE_SHADOW_GPRS is available.