or1k, hooks.S: refactoring exception handlers
Currently, all the handlers are contained within the 256 bytes between the
exception vector offset, which has worked fine so far. It turns out that when
using SMP, we need to put a few more instructions inside each handler to
track the current core and that would cost us some valuable instructions that
could touch in the 256-byte limit.

In an attempt to avoid this, now all the handlers have their own function
which is called from the org directive. Although this ends up wasting some
unnecessary space, this also brings some good things:

- we can use only one jump instruction inside the handler, since we can now
use save/enter as a macro, just as in i386.

- as a side effect of the previous one, it's now possible to debug save/enter
since they are no longer a function.