or1k: [temporarily] waiting cores to initialize
To have a known state, it is interesting, at first, to synchronize the startup
of the cores when in SMP. This allows us to ensure that when the master core
needs to do some IPI between the cores, the other cores are actually initialized
and ready to run.

To achieve this, we have many ways to make the core master wait for slaves, such
as barriers, atomic operations, and so on... The easiest way I found is to
simply delay the unlock and then put a lock operation right after the release.

The changes in smp.c on lines 109 and 110 require a brief explanation: although
strange, this works because the spinlock inside the loop (in smp_init) ensures
that only one core at a time will be released. When the last core is released,
just like the others, it also performs an unlock operation, so the lock on line
109 waits for the last core to release the lock. When core 0 finally manages to
get the lock, we do not really need to continue within a critical section, so
the unlock operation is performed right away.