or1k: moving curr_thread references into per core strucure
As we are moving to SMP support, each thread will be linked to a core, so
we need more than one curr_thread, in this case a curr_thread for each core.
To achieve this, the curr_thread structure has been moved to the per_core
structure and all references to the current thread have been updated to the
new strucutre.