Thread implicit exit call
We add an initialization setup routine in the libc to be able to call pthread_exit implicitly when the main routine of thread returns. To execute it, another function pointer (to this startup routine) is passed to the Kernel: hence we now call the startup routine before the main thread routine (passed by parameters).
- Change functions prototypes.
- Change or1k forge_stack bootstrap (to execute startup instead of main
routine).
- Remove useless pthread_exit() explicit call in the tests.