or1k/i386, bug fix: wrong base pointer while copying stacks
When the kernel forks a new process, a new kernel stack must be created, and
the stack pointer and base pointer are adjusted to point to the appropriate
regions, base in the parent's offset. However, besides the registers, the
frame pointer inside the stack also need to be adjusted, in order to reflect
the new address space.

This commit partially fixes the issue #240, because only the first frame
pointer is adjusted, while the ideal fix would change all the frames in the
stack.