or1k: forging fake stack inside user_mode()
When the core master is serving the syscall execve to a slave, the master
also ends up executing the user_mode() function, that performs the context
switch and executes the process itself, but although the master prepares
the entire environment for the process, who should run it is the slave.

As a way to prevent the master from running the process via user_mode(),
the function has been hacked in a way that all the configured registers are
set directly in the interrupt stack of the slave, so when the slave returns
from the SW-interrupt the leave() function do the context switch and run
the program.