or1k: moving sleeping chain into threads
The current approach of sleeping chains puts the process to sleep while
waiting for a resource. Since the kernel already supports threads, this is
not the expected behaviour since it causes an entire process to sleep if
only one thread is waiting for resources.

This commit, in turn, moves the entire sleeping chain structure to the
thread structure, resulting in a couple of changes into various parts of
the kernle, including unexpected parts, such as scheduler and signal
processing.