or1k: re-scheduling blocking threads
Sometimes it happens that the master core executes some syscalls that are
blocking, for example: I/O functions. When this happens, the kernel saves
both master and slave states on specific stacks and the master jumps to
idle, waiting for the resource.

When the resource comes in, the thread once in WAITING is now ready to
run, but it turns out that the kernel needs to retrieve the same state
again for master *and* slave. To achieve this, the kernel while in
yield_smp() looks for any READY thread that has not returned from
syscall yet (blocking threads) and after scheduling all the 'normal'
threads, the master sends an IPI to the respective slave core and
restores the old state.

If more than one thread is ready to go but waiting for IPI release,
the core master whenever it returns from an IPI handler, it will
search for blocking threads and restore the old context again. This
happens until all threads (for the current process) that are ready
are running.