or1k: changing the timer behaviour
Previously, the code used SPR_TTMR to mask/unmask the timer, which is clearly a
misleading approach, so in an ideal approach (introduced here) the timer always
stays active, and is only masked/unmasked (via SPR_SR_TEE) in some points of the
code.

In addition, it was noted that the 'enable_interrupts()' function is invoked from
some kernel points that may erroneously enable interrupts (as in a kprintf), so
the enable_interrupts() function checks for the current irqlvl before activating
the interrupts.