or1k: using redzone in intstack
When a clock interrupt occurred exactly in the prolog of a function, the kernel
stack (before the interrupt) was being overwritten, since the last stacked values
of intstack overwrote the values stored in the previous stack. As a way of solving
this, we need to ensure there is a gap between the last stacked values and the
intstack, so that there is no overlap of addresses. This gap is guaranteed through
a 128-byte redzone.