i386/or1k: masking interrupts instead of enabling/disabling
So far, at various points in the kernel, the system has enabled and
disabled interrupts at some critical points in the system. It turns
out that by enabling interrupts, we can effectively receive interrupts
on unwanted kernel points, such as kprintf which calls tty_write
that enables interrupts.

In order to avoid this, a safer approach is to use interrupt levels to
mask only unwanted interrupts rather than enabling interrupts. This
ensures that we will not receive interruptions as side effects from
other functions.