i386/or1k: enabling and unmasking ints in one place
When the core master receives a blocking IPI, it needs to unmask all
interrupts and re-enable interrupts. Since this is machine dependent,
the code before in main.c has been moved to utilities.S in a function
called setup_interrupts() which takes care of preparing the
environment to be ready to receive all possible interrupts.