pm: moving curr_proc out of switch_to
When in SMP mode, there is a race condition when the various threads
write to curr_proc. In addition, if the core master receives a clock
interrupt before curr_proc is changed, the core master will have
unpredictable behavior since it will be reading old/invalid data.

As a way of avoiding this, the curr_proc variable is now set by the
master itself prior sending IPIs to the slaves.