or1k: using page_size as a size of initial slave stack
Every core slave needs to have an initial stack. The initial stack of
each core slave was limited to INT_STACK bytes because the core slave
would only need to save the current context while serving an interrupt.

It turns out that the core slave also used it in the interrupt handler,
so the size used exceeded the previously allocated size, which in turn
overwrote global kernel data.

The solution was to increase the initial stack size from all slaves to
8kB, which is size enough for the core slave to work initially.