or1k, bug fix: returning from idle context
When the scheduler chooses another process to execute, it then sends an
IPI_IDLE to all running cores in order to stop them before sending an
IPI_SCHEDULE.

It turns out that when the core slave was given an IPI_SCHEDULE, its
previous context is restored, which in most cases is in the IPI_IDLE
context. When this occurs, the core runs slave_idle again and does
not return from it, since it's in idle.

In order to fix this, a simple solution is to check the status of
the current core in slave_idle, if it's CORE_RUNNING, it's time to
return from idle.