or1k, bug fix: init ipikstack overwriting login page directory
Due to a mis-calculation of the ipikstack start address, the page
directory of the login process was being overwritten and thus
producing misses in itlb in kernel code.

The ipikstack address does not have to start from the end of the
stack and grow to the top (as I had thought), so the ipikstack
address should be the page address returned by the kernel, period.