or1k, bug fix: core master should also do tlb_flush
Previously, in commit #82c0780, tlb flush was introduced in the slave
cores, since the TLB is per core and certain operations in the master
should be reflected in the slave's TLB. However, when a vfault occurs
on a first stack access (demand zero), the entry in the page directory
is refreshed and the stack is cleared via kmemset. When this occurs,
since tlb_flush has been programmed to run only on the slave, the
master has an old version of the TLB data, which in turn partially
corrupts kernel memory and produces unpredictable behavior.

Therefore, this commit starts performing tlb_flush in both cores to
keep memory consistent at all times.