or1k, bug fix: resetting the exception handler right after an exception
When an exception occurs (pfault or vfault) in the kernel context while
handling a syscall from another core, the 'pre'-pfault/vfault functions
appropriately choose which intstack to be used based on the current
value of 'exception_handler'. However, the value of exception_handler
was not being cleaned right after its use, which caused the kernel to
retrieve the wrong intstack based on the old (still saved) value of
exception_handler.

Therefore, the straightforward solution consists of resetting the value
within the IPI handler immediately after handling the exception.