or1k: introduced delayed IPI
At times it's possible for the core master to receive an IPI at the
very moment it's to schedule another process (within yield_smp()).
Since IPIs depend on the context of the current process, it would
not be possible (within what we have today) to handle IPIs of
different processes at the same time, so... we have to ensure that
all IPIs are dealt before a process change.

To ensure that this will happen, some changes have been made to
the kernel:

a) created a struct that saves the basic data of an IPI and is
linked to the thread that trigger it.

b) changes in the scheduler to allow all threads that have pending
IPIs to be checked, and if so, handle them.

With this, it's possible to notice that there are waiting IPIs and
treat them even if a software interrupt is not triggered, and at
different times provided that while running the same process.