or1k, bug fix: IPI_IDLE doesn't need to change thread state
A blocking process was being woken up and scheduled even in the
THRD_WAITING state, since every time an IPI_IDLE occurred, the
thread's state was set to THRD_READY, even if the thread was
not ready for execution.

This commit, in turn, only removes the statement that changes
the thread state in the IPI handler.