or1k, bug fix: wrong return value in syscall
An incorrect return value was being returned when returning from a
system call, even with the system call being executed correctly. A
careful analysis showed that the master processor was saving the
return value in the current intstack of the thread being served,
which is not always correct, since the current intstack might not
be the same as the master started to treat (think that the thread
received an IPI_IDLE or IPI_SCHEDULE in the meantime).

To address this problem, a new pointer has been added in the thread
structure that points to the intstack used by the master during
syscall handling. When the master returns from the handler, it gets
the intstack (to save the return value) from that new pointer
instead of using the current intstack, since the thread can have
two or more levels of interrupts to be handled, and therefore
cannot be based on a fixed stack position, I mean, we cannot assume
that the syscall intstack will always be on the first, second... level.