or1k, bug fix: block_sched in wrong time
When the context of a blocking syscall is resumed, before the slave
arrived at leave and make the jump to userspace the quantum had
already finished and the slave had already received an IPI_IDLE from
the master, which in turn scheduled the same process again. It turns
out that because the slave was interrupted before arrives at leave
function, THRD_SYS is still set in the thread flags, which makes the
master incorrectly identify that it needs to do a sched_block instead
of a normal sched, which consequently caused the master to restore to
an invalid context.

The solution I've found was to reset THRD_SYS on switch_to, at the
time the slave resumes from running a blocking syscall, so it's
guaranteeded that even if the slave is interrupted before doing the
jump to userspace, the flag has already been cleared.