or1k, bug fix: release_ipi in wrong thread
Incorrect IPI releases occurred in the return of a blocking syscall due
to a race condition in the curr_thread variable: master uses the current
thread of the per_core structure to set the appropriate release_ipi to
the thread, whereas the curr_thread variable is only then configured
after the slave executes the switch_to function. Once the master
executes faster than the slave, it would terminate the syscall handling
even before the slave executed the switch_to, which would use a wrong
thread.

A way to avoid this type of race condition is the use of next_thread
variable, since it's set up even before the master begins to return to
the blocking IPI context and it points to the target thread that is to
be executed.