or1k, bug fix: wrong return value in syscall
While returning from a syscall to usermode, was observed that the
return value was different from what should be returned. Like the
previous commit, there was also a race condition here with the
variable curr_thread, since the master uses it to retrieve the
appropriate intstack in which it will be used to write the return
value.

Since it's not guaranteed that the variable curr_thread will be
set by the slave before the master uses it, the direct solution
is the use the variable next_thread, since there are no race
conditions on it.