i386: porting threads to x86 arch
Until now the x86 was not compatible with the threads implementation,
since this was made primarily focused only on or1k. Beyond that, the
SMP support introduced some changes that have changed slightly some
low-level stuffs, like interrupts, process scheduling... so, in fact,
the x86 was not even bulding anymore.

Taking this to account, this commit fixes all the issues found in
order to x86 build and run, recovering all its glory, again =).