[libs] E: kernel config constants can now be read outside Rust
Created a .toml file w/ all kernel config constants. This .toml
file can be easily read for a shell script (e.g., run.sh) so that
we have a single source of truth for target configuration. The logic
to do this can be seen in the build.rs file.

Note: chose to use TOML vs. JSON to conserve the comments that
were present in code.

Signed-off-by: danbugs <danilochiarlone@gmail.com>