[kernel] E: Move Identity Mngt out of the Kernel
This commits moves identity management out of the kernel. And places it
to in the Process Management Daemon. Note that the set/get uid/gid
system call family is also provided by the Linux Daemon. In the long
run, we should place it in the process daemon and have the POSIX layer
to forward system calls to it.

Finally, by moving this functionality out of the kernel, we need to
devise another strategy  to verify if a process has the rights to change
capabilities.