[libs] E: Implement Send/Recv Message Protocol with Gateway
In this commit I implement a simple messaging protocol between linuxd
and the gateway. Instead of sending two messages, one with the size and
one with the payload, we batch them together into a single message.

I implement this protocol in the syscomm libraries as this simplifies
the logic in linuxd itself.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>