[libs] F: Add timestamp_message Macro
In this commit I introduce a new macro in the profiler that allows us to
inject a timestamp in a message's payload.

This macro should only be used together with a client code that is aware
of this behaviour. For example, the macro will only work if the original
payload is all zeros.

The macro is gated behind the timestamp-messages feature flag, that can
be enabled with the TIMESTAMP_MSG make variable.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>