[ci] B: Re-use `venv` for Python Lint
In this commit I update the `python-lint` make target to re-use a python
virtual environment instead of creating a new one from scratch.

I noticed that the post-commit hook was adding a noticeable latency to
every commit, and this should address that.

I choose to create the `./venv` in the root directory and
`.gitignore`-it, but I am happy to consider alternatives.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>