[daemons] B: Re-factor Bind Socket Address
In this commit I refactor the bind-sockaddr parameter in linuxd to
explicitly state that it is binding to the user VM(s). This commit is a
preparation for when we move the gateway connection to a bound socket,
too. In that case it will be convenient to differentiate the two.

I am aware that there is a bit of conflicting terminology around the
codebase, but I decided to go for the correct (as of now) name, even
though it may be inconsistent with other parts of the code.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>