[daemons] F: Bind Gateway Socket in linuxd
In this commit I move the gateway socket to `bind` instead of
`connect`-ing to the gateway.

This situation maps better to a real deployment of linuxd, where we
receive input from a client that is sending us requests, rather than
reading it from a server running on the client side.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>