[utils] B: Make nanvixd Use the New linuxd Args
In this commit I modify nanvixd to use the new argument names in linuxd,
as well as the new connection pattern where clients (like nanvixd)
connect to the gateway socket running in linuxd.

The new connection pattern required a few changes around nanvixd, mostly
to make sure we refresh linuxd for each new sandbox instance, and
dropping it during sandbox unload. The dropping bit is very important as
it cleans-up the socket files, necessary to run multiple benchmarks
back-to-back (i.e. with ./scripts/benchmark.py).

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>