[scripts] B: Update test-nanvixd.sh to nanvixd.elf Changes
In this commit I update test-nanvixd.sh to match the new behaviour of
`nanvixd.elf`. In particular, we cannot wait for nanvixd.elf to create a
certain socket file to check if nanvixd has started, because now these
sockets are created together with the sandbox, not at the beginning.

Instead, I probe nanvixd's readiness by checking if the HTTP socket is
listening.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>