[utils] F: Add Warm Start VMM Benchmark
In this commit I introduce a new micro-benchmark similar to the
warm-start one, where we only measure the time spent inside the VMM.

To do so, we connect the user VM to a UNIX stream that behaves like
linuxd. We wait for the app to send us a ReadRequest, and then measure
the time elapsed between sending the ReadResponse, and receiving a
WriteRequest, which is the behaviour of the echo application.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>