[build] B: Require TOOLCHAIN_DIR Env. Var. To Be Set
With the latest toolchain upgrades, it is no longer supported to have
the toolchain in $NANVIX_ROOT/toolchain, instead it is required (by
./scripts/setup/rust.sh) for $TOOLCHAIN_DIR to be out-of-tree. In the
Makefile, however, we still provide a default value for TOOLCHAIN_DIR
that is invalid.

In this commit I remove the default assignment to $TOOLCHAIN_DIR, and
instead require the variable to be set in the calling environment.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>