[libs] B: Fix Reading Long Messages From Gateway
In this commit I fix a bug I introduced when writting ad-hoc methods to
read/write from/to the gateway.

In particular, if the read buffer exceeded the IKC message size, we
dropped the message with an error of "Message too large". This is
incorrect, as we may want to read an arbitrarily long message from the
gateway. Note that the chunking in small pieces happens in linuxd, not
in the syscomm lib.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>