[ci] F: Move to Self-Hosted Runner Syntax
In this commit I move our self-hosted CI to use a self-hosted GitHub
runner, and not SSH into a remote VM for each command.

This has a much more natural integration with the GHA syntax, but means
we need to use supporting workflows as composite workflows in order to
preserve the caller's state.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>