[utils] E: Refactor Nanvixd for a Unified Control-Plane Logic
In this commit I re-factor nanvixd with two main goals:
1. Remove the `-keep-alive` logic, and
2. Separate the creation of a new sandbox, from the interaction with it.

I could split the two in separate commits, but the logic in 1 is very
intertwined across the whole stack, making it difficult.

The main intuition is that now nanvixd acts more as a regular
control-plane that can spawn new sandboxes when it receives a request of
type `NEW`, and it returns a sandbox (user VM) id and a socket addres.

Users can then use the socket address to interact with the VMs
stdin/stdout, and the VM id to `KILL` the VM.

In the process I also simplify the caching logic, removing a lot of
duplicated locking, and I also disentangle the creation of linuxd
instances (one per tenant) and of user VM instances.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>