[microvm] E: Making `try_receive_*` Reusable
This commit edits two existing methods used to receive messages. Since
the snapshotting protocol involves flushing, it is relevant to know
whether the method detected the channel was empty. If that is the case,
then it has been flushed successfully. Changing the signature to return
a `Result<bool>` allows these methods to be reused for flushing.