[utils] F: Make Nanvixd Send Shutdown Messages To Linuxd
In this commit I implement the Nanvixd support for graceful linuxd shutdown.
When Nanvixd receives a SIGINT, it will send a `Shutdown` message over
the control-plane socket that it shares with linuxd.

This commit misses the linuxd handling to be complete, which is blocked
until we can poll over different sockets in linuxd.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>