[build] F: Drop nix Crate
After a comment in a previous PR, I realized that I was using the `nix`
crate (which I had introduced myself) in a way that could be fully
covered by the `libc` crate (which we really need). In fact, I think the
former is a wrapper of the latter.

In this commit I drop the `nix` crate, and use `libc` where needed.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>