[daemons] F: Make Linuxd Use User VM Handle
In this commit I modify linuxd to use a user VM handle instead of the
individual socket streams.

This simplifies the logic, and it also makes it possible to get rid of
the layer of indirection for the gateway threads. Instead, we now give
downstream consumers the actual socket stream.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>