[daemons] E: Make Linuxd Handle Partial Reads
In this commit I extend linuxd to graecfully handle partial reads from
the non-blocking socket. We do so by adding a small buffer together with
the non-blocking socket stream where we can cache partial reads.

The buffer dynamically grows, because it will always be strictly smaller
than a message size.