[utils] E: Refine Control-Plane Acceptance in Nanvixd
In this commit I refine the logic in nanvixd to accept connections from
the control-plane. In particular, I make all instances in the cache
share a listener socket, which they can mutably borrow to accept one
connection with a timeout.

If linuxd/microvm fail to connect within a timeout, we deem the instance
unresponsive and kill it.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>