[utils] B: Clean-Up User VM Sockets In Nanvixd
In this commit I fix a resource leak in Nanvixd where we were not
properly cleaning-up the user VM sockets. Causing crashes in the CI
nodes.

In an ideal world, these patches would not be needed because the
SocketListener trait should remove the file once it is dropped. This
will be the case when we have a complete graceful shutdown protocol.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>