[microvm] B: Tolerate Race When Connecting To System VM
In this commit I address a race condition between the user VM
connecting to the system VM, and the system VM having the user VM socket
bound and ready to accept connections.

This race condition is rare, but I could reproduce locally by increasing
the number of iterations of the cold-start benchmark to 10000. I think
once the performance issues with hyperlight are fixed we should consider
bumping up the number of iterations in the benchmark.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>