[daemons] F: Make Linuxd Poll For Gateway Connections
In this commit I modify linuxd to keep a poll structure monitoring the
gateway listener socket, and accept incoming connections from the
gateway by calling accept_timeout on the listener socket itself.

We need a different polling structure than the one used for the user
VMs because we want to accept connections from the gateway on-demand, as
user VMs connect. On top of that, once we have accepted the connection,
and have the corresponding stream, we don't need the poll anymore as we
want each worker thread to read/write to the stream directly.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>