[daemons] F: Linuxd Mangages Multiple User VMs
This commit is the last in the series towards getting linuxd to support
managing multiple user VMs. Thanks to the work in previous commits, this
commit is relatively small.

I move the main control structures to be maps indexed by connection ids,
and I add a helper method to accept a new connection. The main diff is
just moving the logic to accept the gateway connection from the main
method to this helper.

I choose a slab for the conn_id -> User VM map because it offers
convenient methods to get the next free entry, a common operation at
runtime. However, for the conn_id -> worker threads map I use a HashMap
because we need to be able to insert/remove elements by key.

The slab is what drives state changes: when user VMs accept/drop
connections we add/remove entries. The worker-thread map reacts in
accordance.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>