[daemons] B: Drain Sockets On Every User VM Poll Notification
In this commit I address a bug that I introduced when making linuxd
poll, and that only manifested when changing the behaviour of the micro
VM.

The bug is relatively simple: when we get a poll notification we must
make sure to drain the corresponding socket until it returns a
WouldBlock. We were not doing that for the user VM connections, so
whenever we had two messages buffered, we would only process the first
one.

This buffering is a timing issue and somewhat rare, but was much more
common in the new micro VM architecture.

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>