[microvm] F: Interrupt vCPU With Signal
In this commit I implement support to interrupt the vCPU thread by
sending it a SIGUSR1 signal.

According to the KVM docs [1], sending a SIGUSR1 to the vCPU thread will
cause KVM_RUN to return with -EINTR. For the time being we treat an
interrupt as an outside request to halt execution, but we can extend this
functionality in the future.

[1] https://www.kernel.org/doc/html/v5.8/virt/kvm/api.html#kvm-set-signal-mask

Signed-off-by: Carlos Segarra <carlos@carlossegarra.com>